test(html): fold the insertion-mode cases into the parser unit suite
They belong next to the rest of the parseHtml unit tests rather than in a
suite of their own; no case, input or snapshot changed.
